// src/common/enums.ts
export enum GalleryItemTypeEnum {
  thumbnail = 'thumbnail',
  product = 'product',
  zoom = 'zoom',
  superZoom = 'superZoom',
  styleSwatch = 'styleSwatch',
}

export enum ProductTypeEnum {
  single = 'CosmeticSingleVariantProduct',
  color = 'ColorVariantProduct',
  smell = 'SmellVariantProduct'
}

export enum ConfigurableOptionValueLabelEnum {
  color = 'color',
  smell = 'smell',
  colorFacets = 'colorFacets'
}